pyproject.toml
[tool.poetry]
name = "css2-llm-integration"
version = "0.1.0"
description = ""
authors = ["Navrin <navman360@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
openai = "^1.14.2"
streamlit = "^1.32.2"
watchdog = "^4.0.0"
#psycopg2-binary = "^2.9.9"
simplejson = "^3.19.2"
ariadne = "^0.23.0"
ariadne-graphql-modules = "^0.8.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}
pydantic = "^2.7.1"
streamlit-modal = "^0.1.2"
termcolor = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
